combine header and implementation together according to google c++ style guide
