+++ theta_dup/BUILD
@@ -3,9 +3,12 @@ load("@rules_cc//cc:defs.bzl", "cc_library")
 cc_library(
     name = "theta_dup",
     hdrs = [
-        "include/theta_sketch_dup.hpp",
-        "include/theta_sketch_dup_impl.hpp",
-        "include/utils.hpp",
+        "include/theta_sketch_dup.h",
+        "include/utils.h",
+    ],
+    copts = [
+        "-Ithird_party/incubator-datasketches-cpp/theta",
+        "-Ithird_party/incubator-datasketches-cpp/common",
     ],
     deps = [
         "//third_party/incubator-datasketches-cpp:theta",
